Create kl_divergV3.py
Here's an updated version of kl_divergence.py that allows for computing the Kullback-Leibler divergence for multiple files and supports comparing different file types such as text files and PDFs:

In this updated script, the user is prompted to enter a directory to scan for files. The script will then list all the supported files (text files and PDFs) in that directory. The user can select the files to compare by entering the corresponding numbers (separated by commas). The selected files are then preprocessed accordingly, and the Kullback-Leibler divergence is calculated for each pair of selected files.

Please note that in order to support PDF file processing, the PyPDF2 library is used. You can install it by running pip install PyPDF2 in your terminal.

The filename suggestion for this script remains kl_divergence.py.