Create bulk_simularity_ranking.py
To make bulk selection easier for the user, we can add a feature that allows them to input ranges of files they want to select. The user can specify the ranges using a format like start-end to select a continuous range, and use commas to separate multiple ranges or individual file numbers.

Let's implement this feature in the code:

With this updated code, the user can input file numbers or ranges separated by commas to select multiple files in bulk. The ranges can be entered in the format start-end (e.g., 1-5) to select a continuous range, and individual file numbers can be entered separately (e.g., 7, 10). The selected files will be used for further calculations and ranking.

As for the file name, let's call the program "bulk_similarity_ranking.py". This name reflects the added functionality for bulk file selection and ranking based on similarity.

The error is related to the missing dependency "PyCryptodome," which is required for the AES algorithm used by PyPDF2 to decrypt the PDF files. To fix this, you need to install the PyCryptodome library. You can do this using the following command in your terminal or command prompt:

Copy code
pip install pycryptodome
After installing PyCryptodome, you can run the code again, and the error should be resolved.